Speed up s3 upload
Previously s3 upload would make a request to gather the md5 for each file which was extreamly slow.
Now it uses the s3.List() function which is extreamly fast at gathering the details in s3.

The s3.Verify() function now takes in a channel of resources from '--source' and a channel of resources from the '--destination'
and finds out which resources are missing in the destination.

This PR also makes sure only the contents of the specified source path makes it into the destination path